refactor: 폴더 분리
- constants, validator, hooks
